Fix/handle rate limit (#204)
Resolving #202 
Added more `ErrorCode` for better clarity.
Changed most errors to `isRetryable` false.
The root case of this is `RATE_LIMIT` which cannot be retried. This is
caused by the account no longer having tokens for the current time
period. It could be hours or days.

Additionally, the `forceSave` was triggering a compaction which could
trigger the `forceSave`. Prevented this by not letting a system message
trigger compaction.